+++ test/clojure_lsp/handlers_test.clj
@@ -120,7 +120,20 @@
 
 (deftest document-symbol
   (h/load-code-and-locs "(ns a) (def bar ::bar) (def ^:m baz 1)")
-  (is (= 1 (count (handlers/document-symbol {:textDocument "file:///a.clj"})))))
+  (h/assert-submaps
+    [{:name "a"
+      :kind :namespace
+      :range {:start {:line 0 :character 0} :end {:line 999999 :character 999999}}
+      :selection-range {:start {:line 0 :character 0} :end {:line 0 :character 5}}
+      :children [{:name "bar"
+                  :kind :variable
+                  :range {:start {:line 0 :character 7} :end {:line 0 :character 15}}
+                  :selection-range {:start {:line 0 :character 12} :end {:line 0 :character 15}}}
+                 {:name "baz"
+                  :kind :variable
+                  :range {:start {:line 0 :character 23} :end {:line 0 :character 35}}
+                  :selection-range {:start {:line 0 :character 32} :end {:line 0 :character 35}}}]}]
+    (handlers/document-symbol {:textDocument "file:///a.clj"})))
 
 (deftest document-highlight
   (let [[bar-start] (h/load-code-and-locs "(ns a) (def |bar ::bar) (def ^:m baz 1)")]